tambahan

+++ PT.CBM/Oprasional/view/VPembelianBBM.php
@@ -670,7 +670,7 @@ aria-hidden="true">
 function sum() {
   var banyak_barang = document.getElementById('qty').value;
   var harga = document.getElementById('harga').value;
-  var result = parseInt(banyak_barang) * parseInt(harga);
+  var result = parseFloat(banyak_barang) * parseFloat(harga);
   if (!isNaN(result)) {
    document.getElementById('jumlah').value = result;
  }
